anlambert/requirements.txt
boto3
chardet
click
humanize
ndjson
pyspark
python-keycloak
requests
swh-model
swh-storage
